requirements.txt
asgiref==3.7.2
attrs==23.2.0
dj-rest-auth==5.0.2
Django==4.2.8
django-cleanup==8.1.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.2.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
